+++ isort/parse.py
@@ -55,7 +55,7 @@ def import_type(line: str, config: Config = DEFAULT_CONFIG) -> Optional[str]:
     """If the current line is an import line it will return its type (from or straight)"""
     if config.honor_noqa and line.lower().rstrip().endswith("noqa"):
         return None
-    elif "isort:skip" in line or "isort: skip" in line:
+    elif "isort:skip" in line or "isort: skip" in line or "isort: split" in line:
         return None
     elif line.startswith(("import ", "cimport ")):
         return "straight"